Catches errors when handling input.
This has to be handled differently. this way we loose bytes from the input stream when they happen to be cut off.
It would be better to
a) get complete chunks of data (maybe by using a Socket instead of unblocking read)
b) handle the case that a sequence isn't complete by re-inserting the data. This has limitations as we don't always know the overall length I suppose