changes Isolate data exchange
Now there is no data copying involved. The Terminal isolate gets Heap memory, fills it and sends the pointer to the buffer to the UI thread.
The UiBufferLines class is used to carry that pointer and to do the offset math in order to get the data back.
Usage of those lines is tracked and when the last user releases it then the memory gets freed again.